+++ packages/theme/src/components/BlogItem.vue
@@ -102,7 +102,7 @@ const resultCover = computed(() => {
     <div class="badge-list mobile-visible">
       <span v-show="author" class="split">{{ author }}</span>
       <span class="split">{{ showTime }}</span>
-      <span v-show="tag?.length" class="split">{{ tag?.join(' · ') }}</span>
+      <span v-if="tag?.length" class="split">{{ tag?.join(' · ') }}</span>
     </div>
   </a>
 </template>